Dia2/auth-service/requirements.txt
Django==5.0
djangorestframework==3.15
djangorestframework-simplejwt==5.3
psycopg2-binary==2.9.10
redis==4.5.0
django-cors-headers==4.0
django-redis==6.0
gunicorn
python-decouple
